test-requirements.txt
-c constraints.txt
-r lib/server/test-requirements.txt

testtools
fixtures
flake8
flask_testing==0.6.1
nose>=1.3.7
coverage>=4.0.3
connexion
pytest
pytest-cov

# The purpose of this file is to contain python runtime requirements
# for testing code, e.g., code authored by developers, as opposed to
# automatically generated code. Requirements from automatically generated
# code are stored in lib/server/test-requirements.txt (within the associated
# docker image), and are referenced above.
